# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Job API",
        "description": "API Documentation for Job Management",
        "version": "1.0.0"
    },
    "paths": {
...
        "/job": {
            "get": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Liste des offres d'emploi",
                "operationId": "e4e920b5692c04b53d252964dc9fc7ec",
                "responses": {
                    "200": {
                        "description": "Liste des jobs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Développeur Laravel"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Nous recherchons un développeur Laravel expérimenté"
                                            },
                                            "company": {
                                                "type": "string",
                                                "example": "Tech Corp"
                                            },
                                            "location": {
                                                "type": "string",
                                                "example": "Paris, France"
                                            },
                                            "salary": {
                                                "type": "number",
                                                "example": 50000
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Créer une offre d'emploi",
                "operationId": "558f0a95151905e8ebf9559f478a7759",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "company",
                                    "location"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Développeur Laravel"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Nous recherchons un développeur Laravel expérimenté"
                                    },
                                    "company": {
                                        "type": "string",
                                        "example": "Tech Corp"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Paris, France"
                                    },
                                    "salary": {
                                        "type": "number",
                                        "example": 50000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Job créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Développeur Laravel"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Nous recherchons un développeur Laravel expérimenté"
                                        },
                                        "company": {
                                            "type": "string",
                                            "example": "Tech Corp"
                                        },
                                        "location": {
                                            "type": "string",
                                            "example": "Paris, France"
                                        },
                                        "salary": {
                                            "type": "number",
                                            "example": 50000
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/job/{id}": {
            "get": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Obtenir les détails d'une offre d'emploi",
                "operationId": "7f9956707d2c653f25fe597621cc4084",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'offre d'emploi",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du job",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Développeur Laravel"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Nous recherchons un développeur Laravel expérimenté"
                                        },
                                        "company": {
                                            "type": "string",
                                            "example": "Tech Corp"
                                        },
                                        "location": {
                                            "type": "string",
                                            "example": "Paris, France"
                                        },
                                        "salary": {
                                            "type": "number",
                                            "example": 50000
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Mettre à jour une offre d'emploi",
                "operationId": "40d81df9dbea26e680636b4d03893a21",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'offre d'emploi",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Développeur Fullstack"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Expérience requise en Laravel et React"
                                    },
                                    "company": {
                                        "type": "string",
                                        "example": "Tech Corp"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Marseille, France"
                                    },
                                    "salary": {
                                        "type": "number",
                                        "example": 60000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Job mis à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Job updated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Supprimer une offre d'emploi",
                "operationId": "f06a25e4a0b4ecb3b9521ee4813866f2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'offre d'emploi",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Job supprimé avec succès"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}